sigwinch and other changes

+++ src/ncsh.c
@@ -23,22 +23,27 @@
 #include "ncsh_config.h"
 #include "ncsh_defines.h"
 #include "ncsh_parser.h"
-#include "ncsh_global.h"
 #include "readline/ncsh_readline.h"
-#include "ncsh_types.h"
+#include "readline/ncsh_autocompletions.h"
 #include "vm/ncsh_vm.h"
-#include "ncsh_global.h"
 
 jmp_buf env;
 sig_atomic_t vm_child_pid;
+volatile int sigwinch_caught;
 
-static void ncsh_signal_handler(int signum, siginfo_t* const info, void* const context)
+static void ncsh_signal_handler(int sig, siginfo_t* const info, void* const context)
 {
-    (void)context; // to prevent compiler warnings
-    const __pid_t target = vm_child_pid;
+    (void)context;
+
+    if (sig == SIGWINCH) {
+        sigwinch_caught = 1;
+        return;
+    }
+
+    const pid_t target = (pid_t)vm_child_pid;
 
     if (target != 0 && info->si_pid != target) { // kill child process with pid vm_child_pid
-        if (!kill(target, signum)) {
+        if (!kill(target, sig)) {
             if (write(STDOUT_FILENO, "\n", 1) == -1) { // write is async/signal safe, do not use fflush, putchar, prinft
                 perror(RED "ncsh: Error writing to standard output while processing a signal" RESET);
             }
@@ -58,19 +63,11 @@ static int ncsh_signal_forward(const int signum)
     act.sa_flags = SA_SIGINFO | SA_RESTART;
 
     if (sigaction(signum, &act, NULL)) {
+        perror("ncsh: couldn't set up signal handler");
         return errno;
     }
 
     return 0;
-}
-
-/* ncsh_exit
- * Called on exit to free memory related to the shells lifetime & the shells main loops lifetime.
- */
-void ncsh_exit(struct ncsh_Shell* const restrict shell)
-{
-    ncsh_readline_exit(&shell->input);
-    z_exit(&shell->z_db);
 }
 
 [[nodiscard]]
@@ -85,9 +82,17 @@ char* ncsh_init_arena(struct ncsh_Shell* const restrict shell)
         return NULL;
     }
 
-    shell->arena = (struct ncsh_Arena){ .start = memory, .end = memory + (arena_capacity) };
+    shell->arena = (struct ncsh_Arena){
+        .start = memory,
+        .end = memory + (arena_capacity),
+        .exit = &env
+    };
     char* scratch_memory_start = memory + (arena_capacity + 1);
-    shell->scratch_arena = (struct ncsh_Arena){ .start = scratch_memory_start, .end = scratch_memory_start + (scratch_arena_capacity) };
+    shell->scratch_arena = (struct ncsh_Arena){
+        .start = scratch_memory_start,
+        .end = scratch_memory_start + (scratch_arena_capacity),
+        .exit = &env
+    };
 
     return memory;
 }
@@ -125,7 +130,7 @@ char* ncsh_init(struct ncsh_Shell* const restrict shell)
         return NULL;
     }
 
-    if (ncsh_signal_forward(SIGINT)) {
+    if (ncsh_signal_forward(SIGINT) || ncsh_signal_forward(SIGWINCH)) {
         perror("ncsh: Error setting up signal handlers");
         return NULL;
     }
@@ -217,8 +222,11 @@ int_fast32_t ncsh(void)
         }
         }
 
-        ncsh_readline_history_and_autocompletion_add(&shell.input,
-                                                     &shell.arena);
+        ncsh_history_add(shell.input.buffer, shell.input.pos, &shell.input.history, &shell.arena);
+        ncsh_autocompletions_add(shell.input.buffer,
+                                 shell.input.pos,
+                                 shell.input.autocompletions_tree,
+                                 &shell.arena);
 
       reset:
         memset(shell.input.buffer, '\0', shell.input.max_pos);
@@ -229,7 +237,8 @@ int_fast32_t ncsh(void)
       }
     } else {
     exit:
-      ncsh_exit(&shell);
+      ncsh_readline_exit(&shell.input);
+      z_exit(&shell.z_db);;
       free(memory);
       puts("exit");
     }
